Added User authenication , so  that only the user which has logged in can add, get or delete expenses

+++ controllers/expenseController.js
@@ -3,6 +3,7 @@ const expenses = require("../models/expenses");
 const path = require("path");
 
 const bcrypt = require("bcrypt");
+const { log } = require("console");
 
 exports.getAddexpense = (req, res) => {
   res.sendFile(path.join(__dirname, "../public/Addexpense.html"));
@@ -11,7 +12,8 @@ exports.getAddexpense = (req, res) => {
 exports.postAddexpense = (req, res) => {
   const { expense, description, category } = req.body;
   // console.log(req.body);
-
+  console.log(req.user.id);
+  console.log("iam here in postaddexpense");
   let values;
 
   const createExpense = expenses
@@ -19,6 +21,7 @@ exports.postAddexpense = (req, res) => {
       expense: expense,
       description: description,
       category: category,
+      userId: req.user.id,
     })
     .then((data) => {
       console.log("expense created");
@@ -29,12 +32,15 @@ exports.postAddexpense = (req, res) => {
     .catch((err) => {
       console.log(err);
       console.log("error occured in creating expense");
-      // res.send("an error occurder cannot create expense");
+      res.send("an error occurder cannot create expense");
     });
 };
 
 exports.getExpenses = async (req, res) => {
-  const allExpenses = await expenses.findAll();
+  // console.log(req.user.id);
+  const allExpenses = await expenses.findAll({
+    where: { userId: req.user.id },
+  });
   console.log(allExpenses);
   res.status(200).send(allExpenses);
 };
@@ -42,7 +48,8 @@ exports.getExpenses = async (req, res) => {
 exports.deleteExpense = (req, res) => {
   console.log(req.params);
   const { id } = req.params;
+  console.log(req.user.id);
   console.log(id);
-  const count = expenses.destroy({ where: { id } });
+  const count = expenses.destroy({ where: { id: id, userId: req.user.id } });
   res.send("deleted successfully");
 };
